Initial version of the TEALS Minecraft Project
These files represent the publicly-visible portion of the TEALS
Minecraft modding project, without solutions, and stripped of some early
large zipped releases that were committed to the repo.

Instructors can find the solutions to the labs in the projects/minecraft
directory of the private APCS instructors' repository on GitHub at
https://github.com/TEALSK12/apcs.

This initial commit is a snapshot of the state of the repo, stripped of
instructor-only material, from 2016-July-9. For a full history of this
project, and if you area an instructor, send email to steve@hollasch.net
or connor@hollasch.net and we can add you to the list of contributors on
the original private TEALSMC repo.