Another attempt at Jekyll; not working.
This one hits a known issue with Jekyll on Windows, which was fixed two
weeks ago, but not yet released. Getting ready to give up on Jekyll, as
it has a lot of complicated requirements to run locally on Windows. In
the end, probably more hassle than just writing HTML.