Serve GitHub pages from master branch
Normally, GitHub pages are served from a special `gh-pages` branch, or a
`docs` directory, or the master branch. This change switches from
serving from the `docs` directory to using the master branch. The main
file is moved into the root of the project, and `docs` has been renamed
to `site`.

This is a simpler mechanism for serving pages, and even allows (I think)
HTML references to components within the project.